move tags thunks to its own file
